feat: Relax ip based rate limit
Multiple users can be on the same IP, so this limit should be larger
than the per-user limit. Setting it to 4x to comfortably allow a full
party of 4 on a LAN.
NOTE: This is still the only non user controlled rate limit, so it
should not be set too high. Will reduce again if people are up to no
good.