fix: Only store minified response
It would be great to be able to store the whole response, but at around
a 200x size increase it is not really worth it.
Should include some more fields that could be relevant in the future in
the minified result so we start storing it as soon as possible.