chore: Add data storage version column to database
This can be useful in the future if we ever make changes to the format
stored in the database. Using this we can conditionally parse the data
into different structures, or exclude incompatible data formats.

Adding it as nullable now, and will start writing to it in the new
deployment. After all old deployments are replaced we will run a
migration that fills the blanks and sets it to be non-nullable.