fix: Don't store duplicate consecutive stats
If you leave the overlay running, it will request your stats every 10
ish minutes. Earlier, this would result in us persisting every copy of
those stats. Now we only persist it if the stats have changed in a
meaningful way (i.e. if any of the fields we store have changed).