fix: Store consecutive duplicate stats if an hour or more apart
Since the overlay requests your own stats at the end of each game, if
you start a new session the next day your first stats won't get
persisted since they are consecutive + duplicate. This breaks session
detection, as the first (baseline) stats of the session won't get
stored.

Not storing duplicate stats is mostly a storage-optimization for the db,
so softening the requirements a bit is totally fine.