fix: Don't return duplicate stats from history endpoint
Our history algorithm divides the timespan into `limit - 1` interval,
and then gets the first stats instance in each of those intervals.
To also have the latest available data, we add the last stats instance
from the last interval. If there was only a single stats instance in the
entire last interval, this would get returned twice. This is now fixed
:^)