fix: Correctly include sessions at the end of the interval
The copilot review identified an issue where the check to see if a
session should be included in the loop was different from the one at the
end. This was due to this check being incorrect to start with, and then
me fixing it only in the loop and not at the end. Added a failing
regression test for this, and then fixed the issue by factoring out the
check into a helper function and using that both places.

This also helped quite a bit with readability, as it allowed splitting
the condition up into two parts that each made a lot of sense.