feat: Add `consecutive` to sessions
This can be false for players who have disabled "Online Game Stats" on
Hypixel, as they may play for a long time without any stat change, and
then re-log and register all the stat changes at once.
This causes extremely short sessions with high stats, which look
nonsensical. This change allows the web app explain this to the user.