chore: Implement window/limit request limiter
This can be used internally in flashlight to limit the rate of outgoing
requests. Even though we have rate limits enforced in our port layer,
multiple different users may request data at the same time, and by the
time the requests to the third party services go through, they may have
bunched up a bit as well.

This implements a window/limit request limiter that only allows a
certain number of requests to go through per time window. The limiter is
conservative, in that it guarantees that any action performed within the
`operation` passed to `Limit()` will not be performed more than `limit`
times within **any** time window of length `window`.

The implementation is mostly ported from prism, with some modifications
to allow canceling requests, and to fit into golang.