chore: Replace minecraftservices endpoints with mojang endpoints
These endpoints have been returning a lot of strange 429s. Switching
back to the username -> uuid endpoint that prism has been using. Also
switching to a uuid -> username endpoint from api.mojang.com as well due
to the same issues.

Hopefully this will resolve the issues. The reason for the switch was
some strange 403s from the mojang API, but those seem to have been
intermittent, and not as consistent as these 429s are being.

https://bugs.mojang.com/browse/WEB/issues/WEB-7591