chore: Rename max operation time -> min operation time
Since we don't want the limiter to eagerly cancel any operation, it
makes more sense to pass in the minimum time the operation is expected
to take. That way we know that if wait + min operation time < deadline,
then there is little chance that the request will succeed, even if we
wait.

Updating the values still remains, this will be done in the following
commit.