fix: Point to image digest when deploying new service
When we were pointing to the `:latest` tag, the service.yaml file would
not change between revisions, which seemed to cause the
`gcloud run services replace` call to not create a new revision.

Switching this to pointing to the specific image that we just built and
pushed to be explicit and to ensure that new revisions are actually
deployed.