chore: Convert player.Experience from float to int
It being a float was a remnant from the fact that the Hypixel API
sometimes returns experience with a decimal separator, or in scientific
notation. To be able to correctly parse it, we need to use a float64 in
the json struct we unmarshal into.
Since we now have separated adapters and domain, we can convert it to an
int64 when storing it there.

Still remaining is converting usage of this value over to int64 as well.