chore: Factor out exp until next star
This was duplicated across the two stars<->exp converters.

Also slightly changed the indicies, as we used to calculate exp to get
to the given star from the one before, but now calculate the exp to get
from the given star to the next one.

Add tests for the new helper, as well as a test asserting that the
expPerPrestige constant gives the same value as summing up all the first
100 levels.