chore: Include progress towards next star in `ExperienceToStars`
Implement the same behavior as in prism, which calculates the fractional
part of the star as the ratio of extra exp to the exp required for the
next star.
This makes the weight of the fractional part different between different
stars, but makes the fractional part accurately represent the progress
towards the next star.