chore: Rework workflow triggers and filters
An issue with the previous workflow setup was that it only ran tests and
staging deploy on pull requests to main, not to other branches. This was
done as the workflow used the pull_request.merged flag to determine
whether to deploy to staging or to production.

Changing this to now run the workflow also on push to main, and using
this as the trigger for deploying to production. This allows us to run
tests and staging deploy on all PRs.

Also removed some checks that reduced the number of times tests and
staging deploys run. Possible to re-add if desired, but should not be a
problem to have them running a bit more often.