chore: Add debug logging to request limiter and urchin api
Have seen some cases of "too many requests to urchin API", which cannot
possibly be the case. My guess is that the context is already cancelled
when entering limiter.Limit, which might give an early return with
`false` - leading to the aforementioned error report.

Adding a bunch of logging so we can figure out what is going on the next
time it happens :^)