chore: Always use logging.FromContext
To ensure we always have the most up-to-date logger, we should fetch it
directly from the context, instead of using one that was previously
stored - potentially before the last update to the logging context.

Since we're mostly logging in error paths anyway, it makes a lot of
sense to not have to fetch the logger every time, only to use it when
something fails.