fix: Make the consecutive check more lenient
This was added to highlight issues with very strange sessions, such as
those caused by "disable online game stats", which can cause a lot of
stats to be registered at once.

The current system can sometimes flag totally normal sessions as not
consecutive, as losses are not registered immediately, and may get
registered at the same time as the next win (or loss). This causes a 2
jump, which we currently don't allow.

Change this to allow 2 jumps to hopefully reduce the rate of these false
positives.